RFTR: WIP- Add RAKE Seed Design
